feat: use quote contexts
This commit moves away from the transformers pipeline which was doing
global processing of orders of various types, causing complicated shared
logic. This PR adds the `QuoteContext` entity which wraps a _top level_
quote request and manages its dependencies and processing.

QuoteContexts have `dependencies` which are other quote requests that
they depend on. These are then deduplciated based on a request key
(basically just info, so tokenin/tokenout but not specifics of config)
and quoted in parallel. After quoting they are `resolved` which is where
the context can perform any mutation or filtering of quotes using the
dependencies specified.

This should simplify code a lot as config-specific logic is isolated
into the context handler rather than run globally in the transformation
pipeline