refactor: cleanup permitData response handling (#144)
This commit cleans up a few things about permitData response handling:

Permits should be returned quote-level, not top-level as they apply to
each quote. It's not clear from the current framing which quote the data
refers to.

The function for fetching permits should not be on the quote interface.
Different quote types will fetch permits in different ways and require
different inputs. This is already clear from the current implementation
where the fn had a weird parameter that was only used by classic and
completely ignored in dutch. This commit removes the field and allows
each quote type to handle it however they please

Move existing allowance fetch to context level. Again this is something
specific to AllowanceTransfer quote types, so should be handled under the classic
context rather than in the global handler

Smaller changes:
- remove _.isEqual in tests as it creates very unreadable logs
- rename permit to permit2 where relevant