feat: add DutchV2Order as a typed wrapper over v1
Essentially v2 inherits all logic of v1 (quoting, parameterization,
etc). but overrides the creation of the order object to inject new v2
fields like cosigner.

In future we may also want to change things like parameterization
specifically for v2 but figured this was faster and relatively clean way to
get to something testable.