fix: portion for synthetic dutch
when requesting dutch only, we build classic request and use its result
to parameterize our portion. however, we do not carry over all the
portion details. This commit updates it to simplify this process and
simply use the portion that already exists in the dutch quote if it is
there